implemented watchdog, uploading files and folder on drive in real time

+++ src/GdriveClient.py
@@ -19,6 +19,16 @@ import sys
 from mimetypes import MimeTypes
 import urllib
 from apiclient import errors
+import Script
+from watchdog.events import FileSystemEventHandler
+from watchdog.events import LoggingEventHandler
+import Queue
+import threading
+import time
+from watchdog.observers import Observer
+from mercurial.util import readfile
+
+#26 May - send all changes from system to drive
 
 
 # try:
@@ -33,9 +43,13 @@ SCOPES = 'https://www.googleapis.com/auth/drive'
 CLIENT_SECRET_FILE = 'client-secret.json'
 APPLICATION_NAME = 'Drive API Python Quickstart'
 home_dir = os.path.expanduser('~');
+folder_path = '/home/piyush/Drive' 
+
     
-
-
+queueLock = threading.Lock()
+workQueue = Queue.Queue(0)
+thread = None
+service  = None
 
 
 
@@ -92,27 +106,7 @@ def get_credentials():
 def getmd5(file):
     return hashlib.md5(open(file, 'rb').read()).hexdigest()
  
-def createDriveFolder(service):
-    file_metadata = {
-        'name' : 'Invoices',
-        'mimeType' : 'application/vnd.google-apps.folder'
-    }
-    file = service.files().create(body=file_metadata,fields='id').execute()
-    print ('File ID: %s' % file.get('id'))
-    
-    # create a folder in system
-    folder_name = 'Drive'
-    path = os.path.join(home_dir, folder_name)
-    folder_path = makeSystemFolder(folder_name, home_dir)
-    dictionary = dict();
-    dictionary['id'] = file.get('id')
-    metadatafile_path = os.path.join(folder_path, '.metadata')
-    metadata_file = open(metadatafile_path, 'r+');
-    serialized = json.dumps(dictionary)
-    metadata_file.write(serialized)
-    metadata_file.flush()
-    
-    
+
 def createreport(service, folder_path):
     report = [];
     print('i am inside create report')
@@ -124,7 +118,7 @@ def createreport(service, folder_path):
         status = json.loads(status_file.read())
     except (ValueError):
          
-        status = None    
+        status = dict()    
     status_file.close()
     
    
@@ -133,19 +127,30 @@ def createreport(service, folder_path):
     dirs = glob.glob(folder_path +'/*')
 #     dirs = os.listdir(path);
     for path_ in dirs:
-        file = open(path_,'r+')
+        if os.path.isdir(path_):
+            continue
+        file = open(path_,'r')
         file_name  = os.path.basename(file.name)
+        if status.has_key(file_name): 
+            print ('sdfdsfdsfdss ' + file_name)
+            drive_id =  status[file_name]['driveid']
+        else:
+            drive_id = None
         if not file.name.startswith('.'):
-            dictionary[file_name] = File(file_name,None, None, status[file_name]['driveid'], getmd5(path_)).__dict__
-            if not status == None:
+            dictionary[file_name] = File(file_name,None, None, drive_id, getmd5(path_)).__dict__
+            if status.has_key(file_name):
                 if not status[file_name]['md5'] == getmd5(path_):
                     print(file_name + " is changed")
-                    report_item = File(file_name, 'upload', 'added', status[file_name]['driveid'])
+                    report_item = File(file_name, 'upload', 'added', drive_id)
                     report.append(report_item.__dict__)
                 else:
                     print(file_name + " is not changed")
-                    report_item = File(file_name, None, 'untouched', status[file_name]['driveid'])
+                    report_item = File(file_name, None, 'untouched', drive_id)
                     report.append(report_item.__dict__)
+            else:
+                print(file_name + " newly added")
+                report_item = File(file_name, None, 'added', drive_id)
+                report.append(report_item.__dict__)
         
     print(json.dumps(report))
     return report
@@ -165,12 +170,15 @@ def sync(service, folder_path):
         print('parent_id' + metadata['id'])
         parent_id = metadata['id']
         status_file_path = os.path.join(folder_path, '.status')
+        if not os.path.exists(status_file_path):
+            open(status_file_path, 'w+').close()
+            
         status_file = open(status_file_path, 'r+')
         status =  None
         try:
             status = json.loads(status_file.read())
         except (ValueError):
-            status = None    
+            status = dict()    
         status_file.close()
         
         report = createreport(service, os.path.join(home_dir, 'Drive') )
@@ -185,7 +193,10 @@ def sync(service, folder_path):
             if not item['driveid']==None:
                 trashDriveFile(service, item['driveid'])
             file_id = uploadfile(service,item['name'], item['driveid'], mime, os.path.join(folder_path, item['name']), parent_id)
-            status[item['name']]['driveid']  = file_id
+            if(status.has_key(item['name'])):
+                status[item['name']]['driveid']  = file_id
+            else:
+                 status[item['name']] = File(item['name'],None, None, file_id, getmd5(os.path.join(folder_path, item['name']))).__dict__   
             
         serialized = json.dumps(status)
         status_file = open(status_file_path, 'w')
@@ -193,19 +204,23 @@ def sync(service, folder_path):
         status_file.flush()
         status_file.close()
             
-def uploadfile(service, name , drive_id, mime, file_path, parent_id):
-#     file_metadata = { 'name' : name }
-#     media = MediaFileUpload(file_path,
-#                         mimetype=mime)
+def uploadfile(service, name = None , drive_id = None, mime = None, file_path=None, parent_id = None):
+#     file_metadata = { 'name' : "pp" }
+#     media = MediaFileUpload('/home/piyush/Drive/piyush',
+#                         mimetype='text/plain')
 #     file = service.files().create(body=file_metadata,
 #                                     media_body=media,
 #                                     fields='id').execute()
 #     print ('File ID: %s' % file.get('id'))  
-    
+    print(parent_id)
+    print('anme ' + name)
+    print('mime' + mime)
+    print('file_path' + file_path)
+#     
     folder_id = parent_id
     file_metadata = {
         'name' : name,
-        'parents': [ folder_id ]
+        'parents': [ parent_id ]
 #          ,'id':drive_id 
     }
     media = MediaFileUpload(file_path,
@@ -214,13 +229,34 @@ def uploadfile(service, name , drive_id, mime, file_path, parent_id):
     file = service.files().create(body=file_metadata,
                                     media_body=media,
                                     fields='id').execute()
-                                    
-    file = service.files().delete()
+                                     
+    
     
     print ('File ID: %s' % file.get('id'))      
-    return file.get('id')    
+    return file.get('id')   
+
+def createDriveFolder(service , name, parent_id):
+    file_metadata = {
+        'name' : name,
+        'mimeType' : 'application/vnd.google-apps.folder',
+        'parents': [parent_id]
+    }
+    file = service.files().create(body=file_metadata,fields='id').execute()
+    print ('File ID: %s' % file.get('id'))
+    return file.get('id')
+    # create a folder in system
+#     folder_name = 'Drive'
+#     path = os.path.join(home_dir, folder_name)
+#     if not (os.path.exists(path)):
+#         folder_path = makeSystemFolder(folder_name, home_dir)
+#     
+     
+
+
 
 def trashDriveFile(service, file_id):
+    if file_id == None:
+        return
     print ('trashing file ' + file_id)
     try:
         return service.files().delete(fileId=file_id).execute()   #trash(fileId=file_id).execute()
@@ -241,8 +277,9 @@ def makeSystemFolder(name, path):
         if(answer == 'No'):
             sys.exit(1)
     # make a metadata file inside this folder, which stores data related to drive
+
     print('making metadata file at' +  path)
-    makeSystemFile('.metadata', path)
+#     makeSystemFile('.report', path)
     return path
     
 def makeSystemFile(name, path):
@@ -261,24 +298,210 @@ class File :
    name = None
    action = None
    status = None # modified, added, deleted, untouched
-   md5 = None   
+   md5 = None
+   is_synced = False
+   is_dir = False   
    
 
-   def __init__(self, name, action = None, status = None, driveid=None, md5 = None):
+   def __init__(self, name=None, action = None, status = None, driveid=None, md5 = None, is_synced=False, is_dir = False):
       self.name = name
       self.action = action
       self.status = status
       self.driveid = driveid
       self.md5 = md5
+      self.is_synced = is_synced
+      self.is_dir = is_dir
       
+
+class LoggingEventHandler(FileSystemEventHandler):
+    """Logs all the events captured."""
+#     makeSystemFile('.report', '/home/piyush/Drive')
+    
+        
+    
+    def on_moved(self, event):
+        print('moved ' + event.src_path + " " + event.dest_path )
+        
+        
+        if not str(event.src_path).find('.goutputstream') == -1:
+            Action(event.src_path, event.dest_path, event.is_directory, 'modified')
+        else:
+            Action(event.src_path, event.dest_path, event.is_directory,'deleted')
+            Action(event.src_path, event.dest_path, event.is_directory,'moved')
+         
+#         
+        
+         
+    
+    def on_created(self, event):
+        print('created ' + event.src_path)
+        Action(event.src_path, None, event.is_directory,'created') 
+
+        
+    def on_deleted(self, event):
+        print('deleted ' + event.src_path)
+        Action(event.src_path, None, event.is_directory,'deleted')  
+        
+        
+    
+    def on_modified(self, event):
+        return
+#         print('modified ' + event.src_path)
+#         if not event.is_directory:
+#             Action(event, 'modified')
+
+
+class Action:
+    
+            
+    def __init__(self, src_path, dest_path, is_dir, status):
+        
+        print('Action is ' + status)
+        file_path = None
+        if status == 'moved' or status == 'modified':
+            file_path = dest_path
+        else:
+            file_path = src_path
+            
+        if ((os.path.basename(file_path)).startswith('.')):
+            return
+        report = dict()
+        report = readFile('/home/piyush/Drive/.report')
+        
+        file = File().__dict__
+        if report.has_key(file_path):
+            file = report[file_path]
+        report[file_path] = File(file_path, 'upload', status, file['driveid'] , file['md5'], False,is_dir).__dict__
+        writeFile('/home/piyush/Drive/.report', json.dumps(report))
+        
+        
+        #change in drive
+        print('here')
+#         queueLock.acquire()
+        print('here2')
+        workQueue.put(report[file_path])
+#         queueLock.release()
+        
+        
+           
+        
+
+def readFile(path):
+    if not os.path.exists(path):
+        report_file = open(path, 'w+').close()
+    report_file = open(path, 'r')
+    try:
+        data = json.loads(report_file.read())
+    except ValueError, e:
+        print(str(e))
+        data = dict()  
+    report_file.close()    
+    return data
+
+def writeFile(path, data):
+    report_file = open(path, 'w+')
+    report_file.write(data)
+    report_file.flush()
+    report_file.close()    
     
 
+
+    
+class myThread (threading.Thread):
+   def __init__(self, threadID, name, q):
+      threading.Thread.__init__(self)
+      self.threadID = threadID
+      self.name = name
+      self.q = q
+   def run(self):
+      print ("Starting " + self.name)
+      process_data(self.name, self.q)
+      print ("Exiting " + self.name)
+
+def process_data(threadName, q):
+   while True:
+#         queueLock.acquire()
+        if not workQueue.empty():
+            file = q.get()
+            print('thread performing ' + file['status'] )
+            report = dict()
+            report = readFile('/home/piyush/Drive/.report')
+            
+#             send file to drive
+            parent_id = getParentid(report, file['name'])
+            if parent_id == None:
+                print('parent_id not found')
+#                 queueLock.release()
+                
+                continue
+            mime = MimeTypes()
+            url = urllib.pathname2url(file['name'])
+            mime_type = mime.guess_type(url)
+            mime = mime_type[0]
+            if mime_type[0]== None:
+                mime = 'text/plain'
+            if file['status'] == 'created' and file['is_synced'] == False:
+                if file['is_dir'] == False:
+                    file_id = uploadfile(service, os.path.basename(file['name']), None, mime, file['name'],parent_id) 
+                else:
+                    file_id = createDriveFolder(service, os.path.basename(file['name']), parent_id)
+                
+                report[file['name']]['is_synced'] = True
+                report[file['name']]['driveid'] = file_id
+            elif file['status'] == 'deleted' and file['is_synced'] == False:
+                trashDriveFile(service, file['driveid'])   
+                report[file['name']]['is_synced'] = True
+            elif file['status'] == 'moved' and file['is_synced'] == False:
+                if file['is_dir'] == False:
+                    file_id = uploadfile(service, os.path.basename(file['name']), None, mime, file['name'],parent_id) 
+                else:
+                    file_id = createDriveFolder(service, os.path.basename(file['name']), parent_id)
+                
+                report[file['name']]['is_synced'] = True
+                report[file['name']]['driveid'] = file_id
+            elif file['status'] == 'modified' and file['is_synced'] == False:
+                trashDriveFile(service, file['driveid']) 
+                file_id = uploadfile(service, os.path.basename(file['name']), None, mime, file['name'],parent_id) 
+                report[file['name']]['is_synced'] = True
+                report[file['name']]['driveid'] = file_id
+            
+            writeFile('/home/piyush/Drive/.report', json.dumps(report))
+            
+#             queueLock.release()
+            print ("%s processing %s" % (threadName, file['name'] +' ' +file['status']))
+#         else:
+#             queueLock.release()
+        time.sleep(1)      
+
+      
+      
+def getParentid(report, path):
+    parent_path = os.path.abspath(os.path.join(path, os.pardir));
+    print('parnet_path ' + parent_path)
+    if str(parent_path).find(folder_path) > -1:
+        print('parent path is valid')
+        if not report.has_key(parent_path):
+            print('created ' + parent_path)
+#            Action(parent_path, None, True,'created')
+            return None
+        else:    
+            parent_drive_id = report[parent_path]['driveid']
+            if parent_drive_id == None:
+                return None
+                
+            else:
+                print("pdi " + parent_drive_id)
+                return parent_drive_id
+             
+    
+    return None
 
 
 def main():
 
      credentials = get_credentials()
      http = credentials.authorize(httplib2.Http())
+     global service
      service = discovery.build('drive', 'v3', http=http)
 #     
 #     results = service.files().list(
@@ -305,6 +528,13 @@ def main():
          createreport(service, os.path.join(home_dir, 'Drive')) 
      if options.sync:
          sync(service, os.path.join(home_dir, 'Drive')) 
+         
+      
+     report = dict()
+     report = readFile('/home/piyush/Drive/.report')  
+     report['/home/piyush/Drive'] = File('/home/piyush/Drive', 'upload', 'created', '0BxW128bZHQogWWVRb3l3eXg1RVk' , None, True, True).__dict__
+     writeFile('/home/piyush/Drive/.report', json.dumps(report))    
+         
 
 if __name__ == '__main__':
     parser = optparse.OptionParser()
@@ -313,6 +543,24 @@ if __name__ == '__main__':
     parser.add_option('-F', '--file',action="store_true", dest='file', help='upload file')
     parser.add_option('-r', '--report',action="store_true", dest='report', help='create report')
     parser.add_option('-s', '--sync',action="store_true", dest='sync', help='sync folder')
+    
     (options, args) = parser.parse_args()
-    print (options.folder)
+#     print (os.path.abspath(os.path.join('/home/piyush/Drive', os.pardir)))
+    
     main()
+    path = folder_path
+    event_handler = LoggingEventHandler()
+    observer = Observer()
+    observer.schedule(event_handler, path, recursive=True)
+    observer.start()
+    thread = myThread(1, "Sync_guy", workQueue)
+    thread.start()
+    try:
+        while True:
+            time.sleep(1)
+    except KeyboardInterrupt:
+        observer.stop()
+    observer.join()
+    
+    
+
